Move link to a separate module.
